requirements.txt
streamlit>=1.30.0
python-pptx>=0.6.23
Pillow>=10.0.0
requests>=2.31.0
duckduckgo-search>=6.0
pytesseract>=0.3.10
fpdf2>=2.7.0
